# docs/openapi3.json
{
  "components": {
...
    "schemas": {
...
      "github_com_ialekseychuk_my-place_internal_dto.BookingResponse": {
        "properties": {
          "client_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "customer_name": {
            "type": "string"
          },
          "end_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "location_id": {
            "type": "string"
          },
          "service_id": {
            "type": "string"
          },
          "service_name": {
            "type": "string"
          },
          "staff_id": {
            "type": "string"
          },
          "staff_name": {
            "type": "string"
          },
          "start_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "github_com_ialekseychuk_my-place_internal_dto.CreateBookingRequest": {
        "properties": {
          "customer_email": {
            "type": "string"
          },
          "customer_name": {
            "maxLength": 100,
            "minLength": 2,
            "type": "string"
          },
          "customer_phone": {
            "type": "string"
          },
          "location_id": {
            "type": "string"
          },
          "service_id": {
            "type": "string"
          },
          "staff_id": {
            "type": "string"
          },
          "start_at": {
            "type": "string"
          }
        },
        "required": [
          "customer_name",
          "customer_phone",
          "service_id",
          "staff_id",
          "start_at"
        ],
        "type": "object"
      },
...
      "github_com_ialekseychuk_my-place_internal_dto.ErrorResponse": {
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "github_com_ialekseychuk_my-place_internal_dto.SlotResponse": {
        "properties": {
          "end": {
            "type": "string"
          },
          "staff_id": {
            "type": "string"
          },
          "start": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "bearerFormat": "JWT",
        "description": "Bearer token (JWT) for authentication. Format: Bearer \u003ctoken\u003e",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "This is a sample API with OpenAPI documentation.",
    "title": "MyPlace API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/businesses/{businessID}/bookings": {
      "get": {
        "description": "Get all bookings for a business with optional date filtering",
        "parameters": [
          {
            "description": "Business ID",
            "in": "path",
            "name": "businessID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Start date in YYYY-MM-DD format",
            "in": "query",
            "name": "start_date",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "End date in YYYY-MM-DD format",
            "in": "query",
            "name": "end_date",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.BookingResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get bookings",
        "tags": [
          "Booking"
        ]
      },
      "post": {
        "description": "Creates a new booking for a service",
        "parameters": [
          {
            "description": "Business ID",
            "in": "path",
            "name": "businessID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.CreateBookingRequest"
              }
            }
          },
          "description": "Booking object",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.BookingResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Service or staff not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Time slot conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Validation errors"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Create a new booking",
        "tags": [
          "Booking"
        ]
      }
    },
    "/api/v1/businesses/{businessID}/bookings/availability": {
      "get": {
        "description": "Get available time slots for a specific business and day",
        "parameters": [
          {
            "description": "Business ID",
            "in": "path",
            "name": "businessID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Date in YYYY-MM-DD format",
            "in": "query",
            "name": "day",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Staff ID to filter availability",
            "in": "query",
            "name": "staff_id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.SlotResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Business not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get available time slots",
        "tags": [
          "Booking"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}